Move shopping cart to it's own file
Move the shopping cart widget to it's own file in order to clean up
the code base.
